+++ src/hooks/index.ts
@@ -12,4 +12,3 @@ export * from './useNavigation';
 export * from './useMediaQuery';
 export * from './usePrevious';
 export * from './useRecoilLoadCatch';
-export * from './useSubmitDistribution';
