+++ src/lib/gql/queries.ts
@@ -1,7 +1,4 @@
-import { order_by } from './__generated__/zeus';
 import { client } from './client';
-
-import { Awaited } from 'types/shim';
 
 export const getCurrentEpoch = async (
   circle_id: number
@@ -22,30 +19,3 @@ export const getCurrentEpoch = async (
   });
   return currentEpoch;
 };
-
-export const getPreviousDistribution = async (
-  circle_id: number
-): Promise<typeof distributions | undefined> => {
-  const { distributions } = await client.query({
-    distributions: [
-      {
-        order_by: [{ id: order_by.desc }],
-        where: {
-          epoch: { circle_id: { _eq: circle_id } },
-          saved_on_chain: { _eq: true },
-        },
-      },
-      {
-        id: true,
-        vault_id: true,
-        distribution_json: [{}, true],
-      },
-    ],
-  });
-  return distributions;
-};
-
-export type PreviousDistribution = Exclude<
-  Awaited<ReturnType<typeof getPreviousDistribution>>,
-  undefined
->[0];
