+++ src/pages/DistributePage/useSubmitDistribution.ts
@@ -2,18 +2,18 @@ import assert from 'assert';
 
 import { BigNumber, FixedNumber, utils } from 'ethers';
 import { ValueTypes } from 'lib/gql/__generated__/zeus';
-import { PreviousDistribution } from 'lib/gql/queries';
 import { createDistribution } from 'lib/merkle-distributor';
 import { MerkleDistributorInfo } from 'lib/merkle-distributor/parse-balance-map';
 import { encodeCircleId } from 'lib/vaults';
 
 import { useDistributor, useApeSnackbar, useContracts } from 'hooks';
+import { Vault } from 'hooks/gql/useVaults';
+
 import {
   useSaveEpochDistribution,
-  useUpdateDistribution,
-} from 'pages/DistributePage/mutations';
-
-import { Vault } from './gql/useVaults';
+  useMarkDistributionSaved,
+} from './mutations';
+import type { PreviousDistribution } from './queries';
 
 export type SubmitDistribution = {
   amount: number;
@@ -29,8 +29,7 @@ export function useSubmitDistribution() {
   const contracts = useContracts();
   const { uploadEpochRoot } = useDistributor();
   const { mutateAsync } = useSaveEpochDistribution();
-  const { mutateAsync: updateDistributionMutateAsync } =
-    useUpdateDistribution();
+  const { mutateAsync: markSaved } = useMarkDistributionSaved();
   const { apeError, showInfo } = useApeSnackbar();
 
   const submitDistribution = async ({
@@ -122,7 +121,7 @@ export function useSubmitDistribution() {
         utils.hexlify(1)
       );
       showInfo('Saving Distribution...');
-      await updateDistributionMutateAsync(response.id);
+      await markSaved(response.id);
       showInfo('Distribution saved successfully');
       return true;
     } catch (e) {
